pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pointwisepy"
version = "1.0.0"
dependencies = [
    "pointwise-glyph-client",
]